not the last commit

+++ Carga.java
@@ -36,7 +36,7 @@ public class Carga{
      * @param quantidade quantidade de items
      * @return true se inserido corretamente
      */
-    public boolean addLoteB(String cliente, int quantidade, double valorUnitario, int desc, int qntDesc) {
+    public boolean addLoteB(String cliente, int quantidade, double valorUnitario, double desc, int qntDesc) {
         if (estaCheia()) return false;
         carga[pos] = new Lote(cliente, quantidade, valorUnitario, desc, qntDesc);
         pos++;
